NP-30600 moved NIN field into own component
